requirements.txt
apache-airflow-providers-apache-spark
azure-storage-blob
delta-spark
pandas
pyspark
pytest
pyarrow
python-dotenv
requests